+++ src/main/java/walter/duncan/dndwebapi/dtos/EquipmentResponseDto.java
@@ -1,7 +1,7 @@
 package walter.duncan.dndwebapi.dtos;
 
 public final class EquipmentResponseDto extends GameInformationResponseDto {
-    public EquipmentResponseDto(String name, String description, int valueInCopperPieces, int weightInLbs) {
+    public EquipmentResponseDto(String name, String description, int valueInCopperPieces, double weightInLbs) {
         super(name, description, valueInCopperPieces, weightInLbs);
     }
 }
